+++ config/src/lib.rs
@@ -1,6 +1,7 @@
 //! Configuration for Drone, an Embedded Operating System.
 
-#![warn(missing_docs)]
+#![feature(unsafe_block_in_unsafe_fn)]
+#![warn(missing_docs, unsafe_op_in_unsafe_fn)]
 #![warn(clippy::pedantic)]
 #![allow(clippy::missing_errors_doc, clippy::module_name_repetitions, clippy::must_use_candidate)]