+++ src/crates.rs
@@ -5,6 +5,7 @@
 #[derive(Clone, Copy, Debug)]
 pub enum Platform {
     Cortexm,
+    Riscv,
 }
 
 /// Drone register and interrupt binding crates.
@@ -13,6 +14,8 @@ pub enum Platform {
 pub enum Bindings {
     Nrf,
     Stm32,
+    Gd32V,
+    Sifive,
 }
 
 /// Drone Serial Output implementation crates.
@@ -20,6 +23,8 @@ pub enum Bindings {
 #[derive(Clone, Copy, Debug)]
 pub enum Dso {
     Nrf91,
+    Gd32V,
+    Sifive,
 }
 
 impl Platform {
@@ -27,6 +32,7 @@ impl Platform {
     pub fn name(self) -> &'static str {
         match self {
             Self::Cortexm => "cortexm",
+            Self::Riscv => "riscv",
         }
     }
 
@@ -34,6 +40,7 @@ impl Platform {
     pub fn flag_name(self) -> &'static str {
         match self {
             Self::Cortexm => "cortexm_core",
+            Self::Riscv => "riscv_core",
         }
     }
 }
@@ -44,6 +51,8 @@ impl Bindings {
         match self {
             Self::Nrf => "nrf",
             Self::Stm32 => "stm32",
+            Self::Gd32V => "gd32v",
+            Self::Sifive => "sifive",
         }
     }
 
@@ -52,6 +61,8 @@ impl Bindings {
         match self {
             Self::Nrf => "nrf_mcu",
             Self::Stm32 => "stm32_mcu",
+            Self::Gd32V => "gd32v_mcu",
+            Self::Sifive => "sifive_mcu",
         }
     }
 }
@@ -61,6 +72,8 @@ impl Dso {
     pub fn name(self) -> &'static str {
         match self {
             Self::Nrf91 => "nrf91",
+            Self::Gd32V => "gd32v",
+            Self::Sifive => "sifive",
         }
     }
 
@@ -119,6 +132,7 @@ impl Dso {
                 "uarte0_ns_txd_maxcnt",
                 "uarte0_ns_txd_ptr",
             ],
+            Self::Gd32V | Self::Sifive => &[],
         }
     }
 }